4차 컬럼 관리 기능 구현(UserRoleEnum Column 추가)

+++ src/main/java/com/example/trelloeaglebrothers/service/ColumnListService.java
@@ -5,6 +5,7 @@ import com.example.trelloeaglebrothers.dto.ColumnListResponseDto;
 import com.example.trelloeaglebrothers.entity.Board;
 import com.example.trelloeaglebrothers.entity.ColumnList;
 import com.example.trelloeaglebrothers.entity.User;
+import com.example.trelloeaglebrothers.entity.UserRoleEnum;
 import com.example.trelloeaglebrothers.repository.BoardRepository;
 import com.example.trelloeaglebrothers.repository.ColumnListRepository;
 import jakarta.transaction.Transactional;
@@ -12,6 +13,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.concurrent.RejectedExecutionException;
 
 @Service
 @RequiredArgsConstructor
@@ -30,25 +32,25 @@ public class ColumnListService {
                 .orElseThrow(() -> new IllegalArgumentException("해당 보드가 존재하지 않습니다."));
 
         //보드 멤버인지 확인
-//        User memberCheck = null;
-//        if (user.getRole().equals(UserRoleEnum.MANAGER) || user.getRole().equals(UserRoleEnum.MEMBER)) {
+        User memberCheck = null;
+        if (user.getRole().equals(UserRoleEnum.MANAGER) || user.getRole().equals(UserRoleEnum.MEMBER)) {
 
-        //보드가 존재 한다면 칼럼 생성
-        ColumnList columnList = new ColumnList(board, requestDto);
+            //보드가 존재 한다면 칼럼 생성
+            ColumnList columnList = new ColumnList(board, requestDto);
 
-        // columnList가 생성 될 때 마다 orderNUm +1 씩 증가
-        //조회를 했을 때 null 이면 1은 넣고
-        //조회를 했을 때 값이 있으면 그 값 플러스 1을 하면 됨
+            // columnList가 생성 될 때 마다 orderNUm +1 씩 증가
+            //조회를 했을 때 null 이면 1은 넣고
+            //조회를 했을 때 값이 있으면 그 값 플러스 1을 하면 됨
 
-        List<ColumnList> columnLists =  columnListRepository.findAllByOrderByOrderNumAsc();
-        Long newOrderNum = columnLists.isEmpty() ? 1 : columnLists.get(columnLists.size() - 1).getOrderNum() + 1;
-        columnList.setOrderNum(newOrderNum);
+            List<ColumnList> columnLists = columnListRepository.findAllByOrderByOrderNumAsc();
+            Long newOrderNum = columnLists.isEmpty() ? 1 : columnLists.get(columnLists.size() - 1).getOrderNum() + 1;
+            columnList.setOrderNum(newOrderNum);
 
-        //저장
-        columnListRepository.save(columnList);
+            //저장
+            columnListRepository.save(columnList);
 
 
-//        } else throw new RejectedExecutionException("접근 권한이 없습니다.");
+        } else throw new RejectedExecutionException("접근 권한이 없습니다.");
 
 
     }
@@ -69,14 +71,14 @@ public class ColumnListService {
                 .orElseThrow(() -> new IllegalArgumentException("해당 칼럼이 존재하지 않습니다."));
 
         //보드 멤버인지 확인
-//        User memberCheck = null;
-//        if (user.getRole().equals(UserRoleEnum.MANAGER) || user.getRole().equals(UserRoleEnum.MEMBER)) {
+        User memberCheck = null;
+        if (user.getRole().equals(UserRoleEnum.MANAGER) || user.getRole().equals(UserRoleEnum.MEMBER)) {
 
 
-        columnList.update(requestDto.getTitle(),board);
-        return new ColumnListResponseDto("컬럼이 변경되었습니다.");
+            columnList.update(requestDto.getTitle(), board);
+            return new ColumnListResponseDto("컬럼이 변경되었습니다.");
 
-//        } else throw new RejectedExecutionException("접근 권한이 없습니다.");
+        } else throw new RejectedExecutionException("접근 권한이 없습니다.");
     }
 
     //칼럼 삭제
@@ -92,12 +94,12 @@ public class ColumnListService {
                 .orElseThrow(() -> new IllegalArgumentException("해당 칼럼이 존재하지 않습니다."));
 
         //보드 멤버인지 확인
-//        User memberCheck = null;
-//        if (user.getRole().equals(UserRoleEnum.MANAGER) || user.getRole().equals(UserRoleEnum.MEMBER)) {
+        User memberCheck = null;
+        if (user.getRole().equals(UserRoleEnum.MANAGER) || user.getRole().equals(UserRoleEnum.MEMBER)) {
 
-        columnListRepository.delete(columnList);
+            columnListRepository.delete(columnList);
 
-//        } else throw new RejectedExecutionException("삭제 권한이 없습니다.");
+        } else throw new RejectedExecutionException("삭제 권한이 없습니다.");
 
 
     }
